Fix missing Running Status support for incoming MIDI on the UART
Running Status is a standard feature of the classic MIDI protocol, which
permits omitting the Status byte when sending a Voice or Mode message
that has the same Status as the preceding message.

Reference: MIDI 1.0 Detailed Specification, v4.2.1 (midi.org)

Many devices make heavy use of this feature and don't work properly with
the KNOT without this implemented. In my testing, this patch fixed MIDI
reception from the following devices:

- Yamaha CBX-K1XG "MIDI sound keyboard" (1995)
- Yamaha QY70 portable sequencer (1997)
- Yamaha PSR-350 arranger keyboard (2001)
- Squier Stratocaster Guitar and Controller (2011)

All were tested with a Yamaha PSS-A50 (2019), a USB-MIDI keyboard/synth.